requirements.txt
python>=3.10
pyyaml
pandas
Mako